move tests to tests folder and modules to modules folder
