+++ modules/array-analysis.js
