+++ modules/strings.js
