+++ tests/strings.test.js
@@ -1,4 +1,4 @@
-import { capitalize, reverseString } from './strings';
+import { capitalize, reverseString } from '../modules/strings';
 
 test('takes a string and returns that string with the first character capitalized.', () => {
   expect(capitalize('hello')).toBe('Hello');
